requirements.txt
aiohttp==3.6.2
altgraph==0.17
async-timeout==3.0.1
attrs==19.3.0
autopep8==1.5
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
future==0.18.2
googletrans==2.4.0
idna==2.9
imageio==2.8.0
keyboard==0.13.5
macholib==1.11
mss==5.0.0
multidict==4.7.5
numpy==1.18.2
olefile==0.46
onnx==1.6.0
onnxruntime==1.2.0
opencv-python-headless==4.2.0.32
pefile==2019.4.18
Pillow==7.0.0
pinyin==0.4.0
protobuf==3.11.3
pycodestyle==2.5.0
pycryptodome==3.8.2
PyInstaller==3.6
PySide2==5.14.2
pytictoc==1.5.0
pywin32==227
pywin32-ctypes==0.2.0
requests==2.23.0
scikit-inpaint==0.1.0
shiboken2==5.14.2
six==1.14.0
typing-extensions==3.7.4.1
urllib3==1.25.8
watchgod==0.6
wincertstore==0.2
yarl==1.4.2
